src/main/java/com/wojtekPi/demo/updaterService/RatioResponse.java
package com.wojtekPi.demo.updaterService;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

import java.util.Map;

@JsonIgnoreProperties(ignoreUnknown = true)
public class RatioResponse {
    private String base;

    public Map<String, String> getRates() {
        return rates;
    }

    public void setRates(Map<String, String> rates) {
        this.rates = rates;
    }

    private Map<String,String> rates;

    public RatioResponse() {
    }

    public String getBase() {
        return base;
    }

    public void setBase(String base) {
        this.base = base;
    }
}
